+++ CHANGES.rst
@@ -5,3 +5,9 @@ Changelog
 -------------------
 
 - Initial release
+
+0.0.2 (2019-08-19)
+-------------------
+
+- RegressionSerialiser.compare now returns a message saying why the comparison failed,
+  or None if the comparison passed, for better failure messages.